requirements.txt
# Core dependencies
crewai==0.28.8
streamlit==1.29.0
python-dotenv==1.0.0
PyYAML==6.0.1

# AI/ML dependencies
openai
langchain>=0.1.10,<0.2.0
langchain-openai>=0.0.2
scikit-learn==1.3.2

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0
isort==5.13.2
autoflake==2.2.1

# Additional utilities
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
setuptools>=65.0.0
thefuzz==0.22.1
openpyxl==3.1.2
